pyproject.toml
[tool.poetry]
name = "lightning-geometric"
version = "0.1.0"
description = "Pytorch Lightning Seed Project with Hydra"
authors = ["Thomas Chaton <thomas.chaton.ai@gmail.com>"]
readme = 'README.md'

[tool.poetry.dependencies]
python = "3.7.8"
torch = "^1.6.0"
pytorch-lightning = "0.10.0"
hydra-core = "^1.0.3"
pytest = "^6.1.1"
wandb = "^0.10.5"
pytest-xdist = "^2.1.0"
scikit-learn = "^0.23.2"
torchvision = "^0.7.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[[tool.poetry.source]]
name = 'private'
url = 'https://pypi.org/project'